split test_common.py into checks and test file. move dataset generation into estimator_checks
